pyproject.toml
[project]
name = "langchain-personal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain-community>=0.3.19",
    "langchain[openai]>=0.3.20",
    "langsmith>=0.3.13",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
]
